Recursion disabled in search for alternative solutions
This is the ONE commit. It solves all the performance problems.
Although I have not tested the correctness by comparison, it should be
equal because the search for alternative solution does the backtrack
starting from elements present in the matrix: those element will likely
to be searched in the next iteration of backtrack.