CMake configuration. Merged debug.h/c to kbest_io_utility.h/c
Using CMake the user can enable or disable the compiling of the debug
prints (for performance). The variable is ported to the code using
an autogenerated file (KBestConfig.h.in).

However the print debug function were outside the public interface,
so I moved the content of utility/debug.h/c inside kbest_io_utility.h/c.